services/auth_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.29
alembic==1.13.1
pydantic==2.8.*
pydantic-settings==2.2.*
python-multipart==0.0.9
passlib[bcrypt]==1.7.*
python-jose==3.3.*
python-dotenv==1.0.*
psycopg2-binary==2.9.10
email-validator==2.1.*
